+++ src/Kambaz/Courses/Piazza/ViewPost/PostBox.tsx
@@ -1,4 +1,3 @@
-
 import "./ViewPost.css";
 
 interface PostBoxProps {
